backend/requirements.txt
asgiref==3.6.0
Babel==2.12.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
coverage==7.2.1
dj-database-url==1.2.0
Django==4.1.5
django-money==3.0.0
django-on-heroku==1.1.2
djangorestframework==3.14.0
Flask==2.1.0
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
krakenex==2.1.0
MarkupSafe==2.1.2
nulltype==2.3.1
numpy==1.24.2
pandas==1.5.3
Pillow==9.3.0
plaid-python==9.1.0
psycopg2-binary==2.9.5
py-moneyed==2.0
pykrakenapi==0.3.1
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
whitenoise==6.4.0
